Adding product specific scripts to be runnable with common scripts
