+++ integration-tests/run-scenario.sh
@@ -17,16 +17,29 @@
 set -e
 set -o xtrace
 
+#Download the common scripts to working directory
+get_cmn_scripts_dwld(){
+git clone https://github.com/wso2-incubator/test-integration-tests-runner.git
+cp test-integration-tests-runner/intg_test_manager.py test-integration-tests-runner/intg_test_constant.py .
+echo "=== Copied common scripts. ==="
+}
+
+get_cmn_scripts_dwld
+
 DIR=$2
 DIR1=integration
 FILE1=${DIR}/infrastructure.properties
 FILE2=${DIR}/testplan-props.properties
 FILE3=run-intg-test.py
-FILE4=configure_product.py
-FILE5=const.py
+FILE4=intg_test_manager.py
+FILE5=intg_test_constant.py
 FILE6=requirements.txt
 FILE7=intg-test-runner.sh
 FILE8=intg-test-runner.bat
+#FILE9=testng.xml
+#FILE10=testng-server-mgt.xml
+#FILE11=$wum_path
+FILE12=prod_test_constant.py
 
 PROP_KEY=keyFileLocation      	 #pem file
 PROP_OS=OS                       #OS name e.g. centos
@@ -145,6 +158,7 @@ if [ "${os}" = "Windows" ]; then
   sshpass -p "${password}" scp -q -o StrictHostKeyChecking=no ${FILE5} ${user}@${host}:${REM_DIR}
   sshpass -p "${password}" scp -q -o StrictHostKeyChecking=no ${FILE6} ${user}@${host}:${REM_DIR}
   sshpass -p "${password}" scp -q -o StrictHostKeyChecking=no ${FILE8} ${user}@${host}:${REM_DIR}
+  sshpass -p "${password}" scp -q -o StrictHostKeyChecking=no ${FILE12} ${user}@${host}:${REM_DIR}
   [ -d ${DIR1} ] && sshpass -p "${password}" scp -q -o StrictHostKeyChecking=no -r ${DIR1} ${user}@${host}:${REM_DIR}
 
   echo "=== Files copied successfully ==="
@@ -169,6 +183,7 @@ else
   scp -o StrictHostKeyChecking=no -i ${key_pem} ${FILE5} ${user}@${host}:${REM_DIR}
   scp -o StrictHostKeyChecking=no -i ${key_pem} ${FILE6} ${user}@${host}:${REM_DIR}
   scp -o StrictHostKeyChecking=no -i ${key_pem} ${FILE7} ${user}@${host}:${REM_DIR}
+  scp -o StrictHostKeyChecking=no -i ${key_pem} ${FILE12} ${user}@${host}:${REM_DIR}
   [ -d ${DIR1} ] && scp -o StrictHostKeyChecking=no -i ${key_pem} -r ${DIR1} ${user}@${host}:${REM_DIR}
 
   echo "=== Files copied successfully ==="
